AIVoifu/voice_conversion/RVC/pyproject.toml
[tool.poetry]
name = "rvc-models"
version = "0.1.0"
description = ""
authors = ["HRNPH <hrnph@protonmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.11"
faiss-cpu = "^1.7.4"
numba = "0.56.4"
numpy = "1.23.5"
scipy = "1.9.3"
librosa = "0.9.2"
llvmlite = "0.39.0"
fairseq = "0.12.2"
gradio = "^3.47.1"
cython = "^3.0.3"
future = ">=0.18.3"
pydub = ">=0.25.1"
soundfile = ">=0.12.1"
ffmpeg-python = ">=0.2.0"
tensorboardx = "^2.6.2.2"
functorch = ">=2.0.0"
jinja2 = ">=3.1.2"
json5 = ">=0.9.11"
markdown = "^3.5"
matplotlib = ">=3.7.1"
matplotlib-inline = ">=0.1.6"
praat-parselmouth = ">=0.4.3"
pillow = ">=9.1.1"
pyworld = ">=0.3.2"
resampy = ">=0.4.2"
scikit-learn = ">=1.2.2"
starlette = ">=0.26.1"
tensorboard = "^2.14.1"
tensorboard-data-server = "^0.7.1"
tensorboard-plugin-wit = "^1.8.1"
torchgen = ">=0.0.1"
tqdm = ">=4.65.0"
tornado = ">=6.2"
werkzeug = ">=2.2.3"
uc-micro-py = ">=1.0.1"
sympy = ">=1.11.1"
tabulate = ">=0.9.0"
pyyaml = ">=6.0"
pyasn1 = ">=0.4.8"
pyasn1-modules = ">=0.2.8"
fsspec = ">=2023.3.0"
absl-py = ">=1.4.0"
audioread = "^3.0.1"
uvicorn = ">=0.21.1"
colorama = ">=0.4.6"
edge-tts = "^6.1.8"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
